[benc/item-flipbook] Further organize ItemFlipbook code
